feat: balance de prueba

+++ src/Repository/Transporte/TteNovedadRepository.php
@@ -244,10 +244,12 @@ class TteNovedadRepository extends ServiceEntityRepository
     {
         foreach ($arrSeleccionados as $arrSeleccionado) {
             $arNovedad = $this->getEntityManager()->getRepository(TteNovedad::class)->find($arrSeleccionado);
-            if ($arNovedad->getEstadoSolucion() == 0) {
-                $this->getEntityManager()->remove($arNovedad);
-            }else {
-                Mensajes::error("No se puede eliminar el registro ya se encuentra solucionado");
+            if($arNovedad) {
+                if ($arNovedad->getEstadoSolucion() == 0) {
+                    $this->getEntityManager()->remove($arNovedad);
+                }else {
+                    Mensajes::error("No se puede eliminar el registro ya se encuentra solucionado");
+                }
             }
         }
         $this->getEntityManager()->flush();
